requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
backoff==2.2.1
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
coverage==7.10.7
cryptography==45.0.4
cssselect2==0.8.0
dnspython==2.7.0
docstring_parser==0.16
ecdsa==0.19.1
email_validator==2.2.0
execnet==2.1.1
fastapi==0.115.12
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
Flask==3.1.0
fonttools==4.58.5
frozenlist==1.7.0
google-ai-generativelanguage==0.6.4
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.92.0
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-cloud-vision==3.4.5
google-crc32c==1.7.1
google-genai==1.14.0
google-generativeai==0.5.4
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpc-google-iam-v1==0.14.2
grpcio==1.72.1
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
langfuse==3.6.2
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.6.4
mysql-connector-python==9.3.0
numpy==2.2.6
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.11.1
packaging==25.0
pdf2image==1.16.3
pdfkit==1.0.0
pillow==11.2.1
pluggy==1.6.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydyf==0.11.0
Pygments==2.19.2
pyparsing==3.2.3
pyphen==0.17.2
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==7.0.0
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==14.1.0
rich-toolkit==0.14.9
rignore==0.6.4
rsa==4.9.1
sentry-sdk==2.34.1
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.42
starlette==0.46.2
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
ujson==5.10.0
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.3
watchfiles==1.1.0
weasyprint==65.1
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
yarl==1.20.1
zipp==3.23.0
zopfli==0.2.3.post1
groq==0.11.0
anthropic==0.42.0
ollama==0.4.4
